app.py
import streamlit as st
import requests
from pptx import Presentation
from pptx.util import Pt
from pptx.dml.color import RGBColor
from pptx.enum.text import PP_ALIGN
import os
import json
import base64
from io import BytesIO
from PIL import Image
import cloudinary
import cloudinary.uploader
import subprocess
import sys

# --- CONFIGURACIÓN DE PÁGINA ---
st.set_page_config(page_title="Generador Provident", layout="wide")

# --- FUNCIONES DE SOPORTE (Basadas en tu código original) ---
def get_libreoffice_path():
    if sys.platform == "darwin": return "/Applications/LibreOffice.app/Contents/MacOS/soffice"
    elif sys.platform == "win32": return "C:\\Program Files\\LibreOffice\\program\\soffice.exe"
    else: return "soffice"

LIBREOFFICE_PATH = get_libreoffice_path()

# Función para procesar la presentación (simplificada para el ejemplo)
def generar_presentacion(datos_cliente, plantilla_path):
    prs = Presentation(plantilla_path)
    # Aquí iría toda tu lógica de reemplazo de texto y fotos
    # ... (omitido por brevedad, pero usaría prs.slides, etc.)
    
    output = BytesIO()
    prs.save(output)
    return output.getvalue()

# --- INTERFAZ DE STREAMLIT ---
st.title("🚀 Generador de Presentaciones Provident")

# Sidebar para configuración
with st.sidebar:
    st.header("Configuración")
    token = st.text_input("Airtable Token", type="password")
    base_id = st.text_input("Base ID")
    table_name = st.text_input("Table Name")
    
    st.divider()
    st.subheader("Cloudinary Config")
    c_name = st.text_input("Cloud Name")
    c_key = st.text_input("API Key")
    c_secret = st.text_input("API Secret", type="password")

# Cuerpo principal
tab1, tab2 = st.tabs(["Generador", "Gestión de Plantillas"])

with tab1:
    st.subheader("Cargar Datos de Airtable")
    if st.button("🔄 Cargar Registros"):
        if token and base_id and table_name:
            # Simulación de carga de datos (Aquí usarías tu lógica de requests)
            st.success("Registros cargados correctamente (Simulación)")
            # st.session_state['data'] = fetch_airtable_data(token, base_id, table_name)
        else:
            st.error("Por favor completa los campos de configuración en la barra lateral.")

    # Selección de Plantilla
    plantilla_subida = st.file_uploader("Sube tu plantilla PowerPoint (.pptx)", type="pptx")

    if plantilla_subida:
        st.info(f"Plantilla seleccionada: {plantilla_subida.name}")
        
        # Botón para procesar
        if st.button("🪄 Generar Presentación"):
            with st.spinner("Procesando..."):
                # Aquí llamarías a tu lógica de procesamiento
                # resultado_binario = generar_presentacion(datos, plantilla_subida)
                
                # Ejemplo de descarga
                st.success("¡Presentación generada con éxito!")
                st.download_button(
                    label="⬇️ Descargar PPTX",
                    data=plantilla_subida, # Aquí iría el resultado_binario
                    file_name="Presentacion_Final.pptx",
                    mime="application/vnd.openxmlformats-officedocument.presentationml.presentation"
                )

with tab2:
    st.subheader("Gestión de Plantillas")
    st.write("Aquí puedes listar y gestionar las plantillas guardadas en la nube.")
    # Implementar lógica de visualización de Cloudinary/Airtable aquí