requirements.txt
streamlit
python-pptx
requests
cloudinary
Pillow